usuarios/requirements.txt
click==6.7
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2-binary==2.7.5
SQLAlchemy==1.2.11
virtualenv==16.0.0
Werkzeug==0.14.1
